Code cleanup
fuq.h was originally implemented in C++. Some of the extra parentheses
were left behind after removing the type casting. These have been
removed.

FUQ_GCC_VERSION was an initial attempt to check the version of GCC used
to compile. It was more trouble than it was worth so the code was
removed, but the undef was forgotten.

Reword #error to state HD memory barrier has not been implemented for
the system. Since it's more likely support is available on a currently
unsupported platform, but it just hasn't been implemented.